level1 (prototype) added. Minor file organization
I made a rough level1 that performs all the necessary functions. Good thing too! Bc I found some glitches

The player can endlessly hop out of the water, I don't know what that is yet

The restart seems to work only for the first level--an error appears when it isn't the first. I don't know where the script is for it either.

This isn't a glitch, but I don't even know how to move images to appear when you go over the shrine and exit portal. Nicole, could you help me?